remove service classes, add common dao
