+++ src/main/java/com/es/phoneshop/dao/Dao.java
@@ -1,8 +1,10 @@
-package com.es.phoneshop.model.product;
+package com.es.phoneshop.dao;
+
+import com.es.phoneshop.model.product.Product;
 
 import java.util.List;
 
-public interface ProductDao {
+public interface Dao {
     Product getProduct(Long id);
     List<Product> findProducts(String query, String order, String sort) ;
     void save(Product product);
