+++ src/main/java/com/es/phoneshop/dao/DaoImpl.java
@@ -1,4 +1,6 @@
-package com.es.phoneshop.model.product;
+package com.es.phoneshop.dao;
+
+import com.es.phoneshop.model.product.Product;
 
 import java.io.Serializable;
 import java.util.ArrayList;
@@ -10,31 +12,13 @@ import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-public class ArrayListProductDao implements ProductDao, Serializable {
+public abstract class DaoImpl implements Dao, Serializable {
     private final static String QUERY_SPLIT = "\\s";
     private final static String SORT_DESCRIPTION = "description";
     private final static String ORDER_DEC = "dec";
 
 
     private final List<Product> productList = new ArrayList<>();
-
-    private static volatile ArrayListProductDao arrayListProductDao = null;
-
-    private static final Object lock = new Object();
-
-    private ArrayListProductDao() {
-    }
-
-    public static ArrayListProductDao getInstance() {
-        if (arrayListProductDao == null) {
-            synchronized (lock) {
-                if (arrayListProductDao == null) {
-                    arrayListProductDao = new ArrayListProductDao();
-                }
-            }
-        }
-        return arrayListProductDao;
-    }
 
     @Override
     public Product getProduct(Long id) {
