+++ src/main/java/com/es/phoneshop/dao/orderDao/OrderDao.java
@@ -1,0 +1,6 @@
+package com.es.phoneshop.dao.orderDao;
+
+import com.es.phoneshop.dao.DaoImpl;
+
+public class OrderDao extends DaoImpl {
+}
