+++ src/main/java/com/es/phoneshop/dao/productDao/ProductDao.java
@@ -1,0 +1,10 @@
+package com.es.phoneshop.dao.productDao;
+
+import com.es.phoneshop.dao.DaoImpl;
+import com.es.phoneshop.model.product.Product;
+
+import java.util.List;
+
+public abstract class ProductDao extends DaoImpl {
+
+}
