+++ src/main/java/com/es/phoneshop/exception/IllegalStockArgumentException.java
@@ -1,4 +1,4 @@
-package com.es.phoneshop.model.exception;
+package com.es.phoneshop.exception;
 
 public class IllegalStockArgumentException extends PhoneshopAppException {
