+++ src/main/java/com/es/phoneshop/model/order/Order.java
@@ -1,0 +1,4 @@
+package com.es.phoneshop.model.order;
+
+public class Order {
+}
