+++ src/main/java/com/es/phoneshop/util/ProductLoader.java
@@ -1,8 +1,8 @@
 package com.es.phoneshop.util;
 
-import com.es.phoneshop.model.product.ArrayListProductDao;
+import com.es.phoneshop.dao.productDao.ArrayListProductDao;
 import com.es.phoneshop.model.product.Product;
-import com.es.phoneshop.model.product.ProductDao;
+import com.es.phoneshop.dao.productDao.ProductDao;
 
 import javax.servlet.http.HttpServletRequest;
