+++ src/main/java/com/es/phoneshop/web/CartPageServlet.java
@@ -1,12 +1,12 @@
 package com.es.phoneshop.web;
 
 import com.es.phoneshop.model.cart.Cart;
-import com.es.phoneshop.model.cart.CartService;
-import com.es.phoneshop.model.cart.CartServiceImpl;
-import com.es.phoneshop.model.exception.IllegalStockArgumentException;
-import com.es.phoneshop.model.product.ArrayListProductDao;
+import com.es.phoneshop.service.CartService;
+import com.es.phoneshop.service.CartServiceImpl;
+import com.es.phoneshop.exception.IllegalStockArgumentException;
+import com.es.phoneshop.dao.productDao.ArrayListProductDao;
 import com.es.phoneshop.model.product.Product;
-import com.es.phoneshop.model.product.ProductDao;
+import com.es.phoneshop.dao.productDao.ProductDao;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
