+++ src/main/java/com/es/phoneshop/web/DeleteCartItemServlet.java
@@ -1,8 +1,8 @@
 package com.es.phoneshop.web;
 
 import com.es.phoneshop.model.cart.Cart;
-import com.es.phoneshop.model.cart.CartService;
-import com.es.phoneshop.model.cart.CartServiceImpl;
+import com.es.phoneshop.service.CartService;
+import com.es.phoneshop.service.CartServiceImpl;
 import com.es.phoneshop.model.product.Product;
 import com.es.phoneshop.util.ProductLoader;
