+++ src/main/java/com/es/phoneshop/web/SampleDataServletContextListener.java
@@ -1,6 +1,6 @@
 package com.es.phoneshop.web;
 
-import com.es.phoneshop.model.product.ArrayListProductDao;
+import com.es.phoneshop.dao.productDao.ArrayListProductDao;
 import com.es.phoneshop.model.product.Product;
 
 import javax.servlet.ServletContextEvent;
