+++ src/main/java/com/es/phoneshop/web/filter/CartFilter.java
@@ -1,7 +1,7 @@
 package com.es.phoneshop.web.filter;
 
-import com.es.phoneshop.model.cart.CartService;
-import com.es.phoneshop.model.cart.CartServiceImpl;
+import com.es.phoneshop.service.CartService;
+import com.es.phoneshop.service.CartServiceImpl;
 
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
