+++ src/main/java/com/es/phoneshop/web/filter/VeiwedProductsFilter.java
@@ -3,8 +3,8 @@ package com.es.phoneshop.web.filter;
 import com.es.phoneshop.model.product.Product;
 import com.es.phoneshop.util.ProductLoader;
 import com.es.phoneshop.model.viewedProduct.ViewedProductList;
-import com.es.phoneshop.model.viewedProduct.ViewedProductService;
-import com.es.phoneshop.model.viewedProduct.ViewedProductServiceImpl;
+import com.es.phoneshop.service.ViewedProductService;
+import com.es.phoneshop.service.ViewedProductServiceImpl;
 
 import javax.servlet.Filter;
 import javax.servlet.FilterChain;
