+++ src/test/java/com/es/phoneshop/web/ProductListPageServletTest.java
@@ -1,7 +1,7 @@
 package com.es.phoneshop.web;
 
 import com.es.phoneshop.model.viewedProduct.ViewedProductList;
-import com.es.phoneshop.model.viewedProduct.ViewedProductService;
+import com.es.phoneshop.service.ViewedProductService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
